+++ src/main/java/vazkii/botania/common/item/relic/ItemFlugelEye.java
@@ -10,15 +10,12 @@
  */
 package vazkii.botania.common.item.relic;
 
-import javax.swing.Icon;
-
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.EnumAction;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.ChunkCoordinates;
-import net.minecraft.util.Vec3;
 import net.minecraft.world.World;
 import vazkii.botania.api.mana.IManaUsingItem;
 import vazkii.botania.api.mana.ManaItemHandler;
@@ -33,7 +30,7 @@ public class ItemFlugelEye extends ItemRelic implements ICoordBoundItem, IManaUs
 	public ItemFlugelEye() {
 		super(LibItemNames.FLUGEL_EYE);
 	}
-	
+
 	private static final String TAG_X = "x";
 	private static final String TAG_Y = "y";
 	private static final String TAG_Z = "z";
@@ -46,7 +43,7 @@ public class ItemFlugelEye extends ItemRelic implements ICoordBoundItem, IManaUs
 				player.swingItem();
 				for(int i = 0; i < 10; i++) {
 					float x1 = (float) (x + Math.random());
-					float y1 = (float) (y + 1);
+					float y1 = y + 1;
 					float z1 = (float) (z + Math.random());
 					Botania.proxy.wispFX(player.worldObj, x1, y1, z1, (float) Math.random(), (float) Math.random(), (float) Math.random(), (float) Math.random() * 0.5F, -0.05F + (float) Math.random() * 0.05F);
 				}
@@ -61,7 +58,7 @@ public class ItemFlugelEye extends ItemRelic implements ICoordBoundItem, IManaUs
 
 		return true;
 	}
-	
+
 	@Override
 	public void onUsingTick(ItemStack stack, EntityPlayer player, int count) {
 		float x = (float) (player.posX - Math.random() * player.width);
@@ -82,16 +79,16 @@ public class ItemFlugelEye extends ItemRelic implements ICoordBoundItem, IManaUs
 		int y = ItemNBTHelper.getInt(stack, TAG_Y, -1);
 		int z = ItemNBTHelper.getInt(stack, TAG_Z, 0);
 		int dim = ItemNBTHelper.getInt(stack, TAG_DIMENSION, 0);
-		
+
 		int cost = (int) (MathHelper.pointDistanceSpace(x + 0.5, y + 0.5, z + 0.5, player.posX, player.posY, player.posZ) * 10);
-		
+
 		if(y > -1 && dim == world.provider.dimensionId && ManaItemHandler.requestManaExact(stack, player, cost, true)) {
 			moveParticlesAndSound(player);
 			if(player instanceof EntityPlayerMP && !world.isRemote)
 				((EntityPlayerMP) player).playerNetServerHandler.setPlayerLocation(x + 0.5, y + 1.6, z + 0.5, player.rotationYaw, player.rotationPitch);
 			moveParticlesAndSound(player);
 		}
-		
+
 		return stack;
 	}
 
@@ -131,7 +128,7 @@ public class ItemFlugelEye extends ItemRelic implements ICoordBoundItem, IManaUs
 
 	/*
 	// Code from the older iteration of this item:
-	
+
 	private static final ResourceLocation glowTexture = new ResourceLocation(LibResources.MISC_GLOW_CYAN);
 
 	private static final int SEGMENTS = 12;
@@ -437,6 +434,6 @@ public class ItemFlugelEye extends ItemRelic implements ICoordBoundItem, IManaUs
 		}
 
 	}
-	*/
+	 */
 
 }
