+++ src/main/java/vazkii/botania/common/item/relic/ItemKingKey.java
@@ -34,15 +34,15 @@ public class ItemKingKey extends ItemRelic implements IManaUsingItem {
 
 	private static final int WEAPON_TYPES = 12;
 	public static IIcon[] weaponIcons;
-	
+
 	public ItemKingKey() {
 		super(LibItemNames.KING_KEY);
 	}
-	
+
 	@Override
 	public void registerIcons(IIconRegister par1IconRegister) {
 		super.registerIcons(par1IconRegister);
-		
+
 		weaponIcons = new IIcon[WEAPON_TYPES];
 		for(int i = 0; i < WEAPON_TYPES; i++)
 			weaponIcons[i] = IconHelper.forName(par1IconRegister, "gateWeapon" + i);
@@ -54,7 +54,7 @@ public class ItemKingKey extends ItemRelic implements IManaUsingItem {
 		setCharging(par1ItemStack, true);
 		return par1ItemStack;
 	}
-	
+
 	@Override
 	public void onPlayerStoppedUsing(ItemStack stack, World world, EntityPlayer player, int time) {
 		int spawned = getWeaponsSpawned(stack);
@@ -63,32 +63,32 @@ public class ItemKingKey extends ItemRelic implements IManaUsingItem {
 			setWeaponsSpawned(stack, 0);
 		}
 	}
-	
+
 	@Override
 	public void onUsingTick(ItemStack stack, EntityPlayer player, int count) {
 		int spawned = getWeaponsSpawned(stack);
-		
+
 		if(count != getMaxItemUseDuration(stack) && spawned < 20 && !player.worldObj.isRemote && ManaItemHandler.requestManaExact(stack, player, 150, true)) {
 			Vector3 look = new Vector3(player.getLookVec());
 			look.y = 0;
 			look.normalize().negate().multiply(2);
 			int div = spawned / 5;
 			int mod = spawned % 5;
-			
-			Vector3 pl = look.copy().add(Vector3.fromEntityCenter(player)).add(0, 1.6, (double) div * 0.1);
-			
+
+			Vector3 pl = look.copy().add(Vector3.fromEntityCenter(player)).add(0, 1.6, div * 0.1);
+
 			Random rand = player.worldObj.rand;
 			Vector3 axis = look.copy().normalize().crossProduct(new Vector3(-1, 0, -1)).normalize();
 			Vector3 axis1 = axis.copy();
 
 			double rot = mod * Math.PI / 4 - Math.PI / 2;
-			
+
 			axis1.multiply(div * 3.5 + 5).rotate(rot, look);
 			if(axis1.y < 0)
 				axis1.y = -axis1.y;
-				
+
 			Vector3 end = pl.copy().add(axis1);
-			
+
 			EntityBabylonWeapon weapon = new EntityBabylonWeapon(player.worldObj, player);
 			weapon.posX = end.x;
 			weapon.posY = end.y;
@@ -97,7 +97,7 @@ public class ItemKingKey extends ItemRelic implements IManaUsingItem {
 			weapon.setVariety(rand.nextInt(WEAPON_TYPES));
 			weapon.setDelay(spawned);
 			weapon.setRotation(MathHelper.wrapAngleTo180_float(-player.rotationYaw + 180));
-			
+
 			player.worldObj.spawnEntityInWorld(weapon);
 			player.worldObj.playSoundAtEntity(weapon, "botania:babylonSpawn", 1F, 1F + player.worldObj.rand.nextFloat() * 3F);
 			setWeaponsSpawned(stack, spawned + 1);
@@ -114,23 +114,23 @@ public class ItemKingKey extends ItemRelic implements IManaUsingItem {
 		return 72000;
 	}
 
-	
+
 	public static boolean isCharging(ItemStack stack) {
 		return ItemNBTHelper.getBoolean(stack, TAG_CHARGING, false);
 	}
-	
+
 	public static int getWeaponsSpawned(ItemStack stack) {
 		return ItemNBTHelper.getInt(stack, TAG_WEAPONS_SPAWNED, 0);
 	}
-	
+
 	public static void setCharging(ItemStack stack, boolean charging) {
 		ItemNBTHelper.setBoolean(stack, TAG_CHARGING, charging);
 	}
-	
+
 	public static void setWeaponsSpawned(ItemStack stack, int count) {
 		ItemNBTHelper.setInt(stack, TAG_WEAPONS_SPAWNED, count);
 	}
-	
+
 	@Override
 	public boolean isFull3D() {
 		return true;
@@ -140,5 +140,5 @@ public class ItemKingKey extends ItemRelic implements IManaUsingItem {
 	public boolean usesMana(ItemStack stack) {
 		return true;
 	}
-	
+
 }
